+++ R/misc.R
@@ -43,9 +43,14 @@ rep.units = function(x, ...) {
 #' parse_unit("kg m-2 s-1")
 #' @details see also \code{demo(cf)} for parsing units in the CF standard name table. Note that \code{parse_unit} currently fails on expressions containing a \code{/}, such as \code{m/s-1}.
 #' @export
-parse_unit = function(str) {
+parse_unit <- function(str) {
+  warning("parse_unit() is deprecated. Please use parse_units(x, implicit_exponents = TRUE) instead")
+  .parse_units_with_implicit_exponents(str)
+}
+
+.parse_units_with_implicit_exponents <- function(str) {
   if (length(grep(c("[*/]"), str)) > 0)
-    stop("parse_unit does not parse unit strings containing `*' or `/'")
+    stop("If 'implicit_exponents = TRUE', strings cannot contain `*' or `/'")
   parse_one = function(str) {
     r <- regexpr("[-0-9]+", str)
     if (r == -1)